+++ src/main/java/com/yukgaejang/inflearnclone/domain/board/dao/UserDao.java
@@ -1,6 +1,6 @@
 package com.yukgaejang.inflearnclone.domain.board.dao;
 
-import com.yukgaejang.inflearnclone.domain.board.domain.User;
+import com.yukgaejang.inflearnclone.domain.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserDao extends JpaRepository<User, Long> {
